🚧 1J49NQ task: split routing and task hotspots
