+++ .agentplane/tasks/202607092208-1J49NQ/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 1J49NQ task: Split routing and task-command hotspots for v0.6.22 [202607092208-1J49NQ]
